+++ app/Http/Controllers/ClassController.php
@@ -2,13 +2,19 @@
 
 namespace App\Http\Controllers;
 
+use App\Classes\ClassHelper;
+use App\Http\Controllers\Approval\ClassApprovalController;
 use Carbon\Carbon;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
 use Inertia\Inertia;
 
 class ClassController extends Controller
 {
+    public $update_class_config = 3;
+    public $delete_class_config = 4;
+
     public function index(Request $request)
     {
         $allowed_centres    =   (object)Inertia::getShared('allowed_centres');
@@ -153,6 +159,47 @@ class ClassController extends Controller
             'end_time'              => 'required',
         ]);
         
+        if(auth()->user()->is_admin == false){
+            $approval_data      =   $request->all();
+            $pending_approval   =   ClassHelper::checkClassPreviousApprovals($request->class_id, $this->update_class_config);
+        
+            if($pending_approval){
+                return redirect(route('programmes'))->with(['type' => 'error', 'message' => 'This programme is on pending approval!']);
+            }
+            $approval_data['class_info']    =   DB::table('classes')
+                                                    ->join('programme_levels','classes.programme_level_id','=','programme_levels.id')
+                                                    ->join('programmes','programme_levels.programme_id','=','programmes.id')
+                                                    ->where('classes.id', $request->class_id)
+                                                    ->select([  'classes.id as class_id','programmes.id as programme_id','programmes.name','programme_levels.id as programme_level_id',
+                                                                'programme_levels.level','classes.centre_id','classes.class_day_id',
+                                                                'programme_levels.class_type_id','classes.class_method_id','classes.start_time','classes.end_time','classes.capacity'
+                                                                ,'classes.status'])->first();
+
+            $approval_data['class_types']   =   DB::table('programme_levels')
+                                                    ->join('class_types','programme_levels.class_type_id','=','class_types.id')
+                                                    ->select(['class_types.id', 'class_types.name'])
+                                                    ->distinct('class_types.id')
+                                                    ->where('programme_levels.programme_id', $approval_data['class_info']->programme_id)
+                                                    ->orderBy('class_types.id')
+                                                    ->get();
+
+            $approval_data['class_levels']  =   DB::table('programme_levels')
+                                                    ->select(['id', 'level'])
+                                                    ->where('programme_id', $approval_data['class_info']->programme_id)
+                                                    ->where('class_type_id', $approval_data['class_info']->class_type_id)
+                                                    ->orderBy('level')
+                                                    ->get();
+
+            $approval_data['programme_list']    =   DB::table('programmes')->get();
+            $approval_data['day_list']          =   DB::table('class_days')->get();
+            $approval_data['type_list']         =   DB::table('class_types')->get();
+            $approval_data['method_list']       =   DB::table('class_methods')->get();
+
+            $approval   =   new ClassApprovalController();
+            $approval->sendClassUpdateRequest($approval_data);
+
+            return redirect(route('classes'))->with(['type' => 'success', 'message' => 'Your request has been sent for approval!']);
+        }
         DB::table('classes')->where('id', $request->class_id)->update([
             'centre_id'             =>  $request->centre_id,
             'programme_level_id'    =>  $request->programme_level_id,
@@ -170,6 +217,28 @@ class ClassController extends Controller
 
     public function destroy($id)
     {
+        /* Check if programme can be deleted */
+        $class_is_deletable   =   ClassHelper::checkClassIsDeletable($id);
+        if(!$class_is_deletable){
+            return redirect(route('classes'))->with(['type' => 'error', 'message' => 'There are students in this class!']);
+        }
+
+        /* Check if user is admin */
+        if(auth()->user()->is_admin == false){
+
+            /* Check if delete request on this programme is still pending */
+            $pending_approval   =   ClassHelper::checkClassPreviousApprovals($id, $this->delete_class_config);
+            if($pending_approval){
+                return redirect(route('classes'))->with(['type' => 'error', 'message' => 'This class is on pending approval!']);
+            }
+
+            /* Send programme delete request */
+            $approval   =   new ClassApprovalController();
+            $approval->sendClassDeleteRequest($id);
+
+            return redirect(route('classes'))->with(['type' => 'success', 'message' => 'Your request has been sent for approval!']);
+        }
+
         DB::table('classes')->where('id', $id)->delete();
 
         return redirect(route('classes'))->with(['type'=>'success', 'message'=>'Class deleted successfully !']);
